# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Riwi Cine API",
    "version": "1.0.0",
    "description": "Documentación generada automáticamente con Swagger para la API de Riwi Cine."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Servidor local versionado (RN-113)"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Legacy sin versionar (compatibilidad)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/functions/movie/{id}": {
      "get": {
        "summary": "Obtener funciones de una película con filtros (formato, fecha, cine)",
        "tags": [
          "Functions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID de la película"
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "type": "string",
              "example": "2D"
            },
            "description": "Formato de proyección (2D, 3D, IMAX, VIP)"
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "type": "string",
              "example": "2026-08-24"
            },
            "description": "Fecha de la función (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "cinemaId",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID del cine o complejo"
          }
        ],
        "responses": {
          "200": {
            "description": "Funciones obtenidas exitosamente"
          },
          "404": {
            "description": "Película no encontrada"
          }
        },
        "description": "### Obtener funciones de una película con filtros\n\nConsulta las funciones disponibles para una película según formato de proyección (2D, 3D, IMAX, VIP), fecha y complejo.\n\n- **Parámetros de ruta:** `:id` (ID de la película).\n- **Query Params:** `format`, `date`, `cinemaId`.\n- **Respuesta:** Devuelve `success` y `data` con el listado de funciones y salas."
      }
    },
    "/functions/{id}/prices": {
      "get": {
        "summary": "Obtener el desglose de precios y promociones de una función",
        "tags": [
          "Functions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID de la función"
          }
        ],
        "responses": {
          "200": {
            "description": "Precios calculados exitosamente"
          },
          "400": {
            "description": "Función inactiva o ya iniciada"
          },
          "404": {
            "description": "Función no encontrada"
          }
        },
        "description": "### Obtener desglose de precios y promociones de una función\n\nCalcula el valor base de la entrada, recargos por sala especializada (IMAX, 3D, VIP) y promociones del día aplicables.\n\n- **Parámetros de ruta:** `:id` (ID de la función).\n- **Respuesta:** Devuelve `success` y `data` con el desglose tarifario."
      }
    },
    "/functions/{id}/seats": {
      "get": {
        "summary": "Obtener el mapa de sillas y disponibilidad en tiempo real de una función (HU-010)",
        "tags": [
          "Functions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID de la función"
          }
        ],
        "responses": {
          "200": {
            "description": "Mapa interactivo de sillas obtenido exitosamente"
          },
          "400": {
            "description": "Función inactiva o ya iniciada"
          },
          "404": {
            "description": "Función no encontrada"
          }
        },
        "description": "### Obtener mapa de sillas y disponibilidad en tiempo real (HU-010)\n\nRepresentación matricial de la sala con estado en vivo de cada silla: disponible, ocupada, bloqueada o reservada.\n\n- **Parámetros de ruta:** `:id` (ID de la función).\n- **Respuesta:** Devuelve `success` y `data` con el layout de la sala y estado de asientos."
      }
    },
    "/functions/{id}": {
      "get": {
        "summary": "Obtener el detalle de una función",
        "tags": [
          "Functions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID de la función"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle de la función obtenido exitosamente"
          },
          "400": {
            "description": "Función inactiva o ya iniciada"
          },
          "404": {
            "description": "Función no encontrada"
          }
        },
        "description": "### Obtener detalle de una función\n\nInformación sobre sala, complejo, película, formato de proyección y horario de inicio.\n\n- **Parámetros de ruta:** `:id` (ID de la función).\n- **Respuesta:** Devuelve `success` y `data` con los datos de la función."
      }
    },
...
  },
...
}